Build Crystal from source for each platform
Use source package instead of official binaries with the objective of
keeping debug symbols and other details useful during development (and
possible in the future enable the interpreter mode).

This process is achieved in multiple stages, first bootstrapping using
existing Alpine's Crystal+Shards version to produce object files that
later are linked into the target executables, which then is combined
with development packages and utilities.

Each Dockerfile uses the following multi-stage sequence:

- stage0: use build platform and existing compiler to cross-compile to target
- stage1: link cross-compiled objects into executables on the target platform
- stage2: prepare directory structure, source code and final executables
- stage3: copy over artifacts and development utilities and dependencies

Note the mentions of _build platform_ and _target platform_, as this
Dockerfile is prepared to cross-compile to different targets using
Docker's BuildKit functionality (Eg. compile for arm64 using amd64
host). This is necessary to prepare for the introduction of
multi-architecture container images.